Plot global profile for a given condition on a given metadata

Plot global profile for a given condition on a given metadata

Usage

plot_global_profile(
  ribo = ribo_adj_annot,
  condition_col = condition_col,
  ech = ech,
  ctrl = ctrl
)

Arguments

ribo

A RiboClass object.

condition_col

condition column for the plot

ech

modality 1 of condition_col

ctrl

modality 2 of condition_col

Value

a ggplot object

Examples

plot_global_profile(ribo, condition_col, ech, ctrl)
#> Error: object 'ribo' not found