Function to generate HTML report using Rmarkdown

Note : if you want to create your own report template, please make sure it has the following header :

— title: "RMSAnalyzer : title of the report" output: html_document

params: library_col: "library" condition_col: "condition" ribo_name: "ribo" project_name: "Unnamed project" —

Usage

report(
  ribo,
  ribo_name,
  rmdfile,
  condition_col,
  library_col,
  project_name,
  output_dir
)

Arguments

ribo

a RiboClass object.

rmdfile

filename of the Rmarkdown template to use.

condition_col

name or index of the column __in metadata__ containing the condition

library_col

name or index of the column __in metadata__ containing the library

project_name

Name of the project to display on the report.

output_dir

Path of output directory