rRMSAnalyzer package dedicated to analysis of RiboMethSeq data

The rRMSAnalyzer package is designed to give easy-to-use functions dedicated to RiboMethSeq (RMS) analyses. All functions revolve around the RiboClass, a S3 class containing both RMS data and samples metadata.

Details

The main functions are (in a standard workflow chronological order):

load_ribodata - import data and create a RiboClass. Objects of this class are used by all other functions

compute_cscore - compute C-score for all genomic positions in each sample

adjust_bias - adjust technical biases of your RiboMethSeq data with ComBat-seq

extract_data - get a dataframe of C-score for all samples and all positions in the RiboClass

Plots, e.g.: plot_pca; plot_coa or plot_heatmap.

To know how to start using this package, have a look at the vignette at vignette("rRMSAnalyzer"). It will give an example for most commands and a workflow.

The code can be viewed at the GitHub repository :

https://github.com/RibosomeCRCL/rRMSAnalyzer

See also

Author

Théo COMBE, Hermes PARAQINDES, Janice KIELBASSA